+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.2.0
+
+* Decrease heading level by 1 since GitBoot put title to H1.
+* Remove front matter but keep author and update date in content.
+
 # 1.1.0
 
 * Change build tool from gulp to npm script